+++ src/spline_functions/build_f2py_modules.sh
@@ -1,0 +1,1 @@
+python3 -m numpy.f2py -c -m bspline_basis_functions_pmod bspline_basis_functions.f90
